Use positions relative to track size of crop automation
The positions are 0..1 relative to the dimensions of the current track.
Use SPDX-License tag.

+++ cinelerra/cropauto.h
@@ -26,10 +26,10 @@ public:
 	size_t get_size();
 	void dump(int indent = 0);
 
-	int left;
-	int right;
-	int top;
-	int bottom;
+	double left;
+	double right;
+	double top;
+	double bottom;
 	int apply_before_plugins;
 };
